Combine the tests into 1 file
