+++ e2e_playwright/st_effects_with_fragment_interactions.py
@@ -22,5 +22,12 @@ def my_fragment() -> None:
 
 my_fragment()
 
-if st.button("Balloons"):
-    st.balloons()
+col1, col2 = st.columns(2)
+
+with col1:
+    if st.button("Balloons"):
+        st.balloons()
+
+with col2:
+    if st.button("Snow"):
+        st.snow()
